# backend/openapi.yaml
openapi: 3.0.0
info:
  title: MealDrop API
  version: 1.0.0
components:
  schemas:
...
    Restaurant:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        address:
          type: string
        image:
          type: string
        owner:
          type: string
          format: uuid
      required:
        - name
        - description
        - address
        - image
        - owner
      additionalProperties: false
paths:
...
  /api/v1/restaurantPartner/create-restaurant:
    post:
      summary: Create a new restaurant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Restaurant'
      responses:
        '201':
          description: Restaurant created successfully
        '400':
          description: Bad request, validation errors
        '500':
          description: Internal server error
  /api/v1/restaurantPartner/list-restaurants:
    get:
      summary: List all restaurants
      responses:
        '200':
          description: Successful response
        '500':
          description: Internal server error
...
  /api/v1/restaurantPartner/get-restaurant-orders/{restaurantId}:
    get:
      summary: Get all customer orders for a specific restaurant
      parameters:
        - name: restaurantId
          in: path
          required: true
          description: The ID of the restaurant
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
        '500':
          description: Internal server error
  /api/v1/restaurantPartner/list-meals/{restaurantId}:
    get:
      summary: List meals for a specific restaurant
      parameters:
        - name: restaurantId
          in: path
          required: true
          description: The ID of the restaurant
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful response
        '500':
          description: Internal server error